These are my starting files as of 5/10/2019.
These are files that I have worked on for about 25 hours. After much trial and error of design, this is what I ended with before I start to finalize the design and add functionality.

Also, note that there is limited functionality. The reason behind this is to get a grasp on what I want from my CSS before I start into the project too much. This project may be redone using a framework such as bootstrap