Test Gallery Changes
I added a decent amount of changes to the galleryTester.html file. They were implemented in two different ways. The first way to view pictures can be done on the site, but to see them up close, you simply click on them and a modal will appear. The purpose of this is to offer different ways to view the picture.